+++ devolutions-gateway/src/lib.rs
@@ -16,7 +16,7 @@ pub mod api;
 pub mod cli;
 pub mod config;
 pub mod credential;
-pub mod credential_injection_kdc;
+pub mod credential_injection;
 pub mod extract;
 pub mod generic_client;
 pub mod http;
@@ -62,7 +62,8 @@ pub struct DgwState {
     pub shutdown_signal: devolutions_gateway_task::ShutdownSignal,
     pub recordings: recording::RecordingMessageSender,
     pub job_queue_handle: job_queue::JobQueueHandle,
-    pub credentials: credential_injection_kdc::CredentialService,
+    pub provisioning: provisioning::ProvisioningStore,
+    pub synthetic_kdc_registry: credential_injection::SyntheticKdcRegistry,
     pub monitoring_state: Arc<network_monitor::State>,
     pub traffic_audit_handle: traffic_audit::TrafficAuditHandle,
     pub agent_tunnel_handle: Option<Arc<agent_tunnel::AgentTunnelHandle>>,
@@ -90,7 +91,8 @@ impl DgwState {
         let (shutdown_handle, shutdown_signal) = devolutions_gateway_task::ShutdownHandle::new();
         let (job_queue_handle, job_queue_rx) = job_queue::JobQueueHandle::new();
         let (traffic_audit_handle, traffic_audit_rx) = traffic_audit::TrafficAuditHandle::new();
-        let credentials = credential_injection_kdc::CredentialService::new(conf_handle.clone());
+        let provisioning = provisioning::ProvisioningStore::new();
+        let synthetic_kdc_registry = credential_injection::SyntheticKdcRegistry::new();
         let monitoring_state = Arc::new(network_monitor::State::new(Arc::new(MockMonitorsCache))?);
 
         let state = Self {
@@ -103,7 +105,8 @@ impl DgwState {
             recordings: recording_manager_handle,
             job_queue_handle,
             traffic_audit_handle,
-            credentials,
+            provisioning,
+            synthetic_kdc_registry,
             monitoring_state,
             agent_tunnel_handle: None,
         };
