+++ devolutions-gateway/src/listener.rs
@@ -158,7 +158,8 @@ async fn handle_tcp_peer(stream: TcpStream, state: DgwState, peer_addr: SocketAd
                 .sessions(state.sessions)
                 .subscriber_tx(state.subscriber_tx)
                 .active_recordings(state.recordings.active_recordings)
-                .credentials(state.credentials)
+                .provisioning(state.provisioning)
+                .synthetic_kdc_registry(state.synthetic_kdc_registry)
                 .agent_tunnel_handle(state.agent_tunnel_handle)
                 .build()
                 .serve()
