+++ devolutions-gateway/src/ngrok.rs
@@ -237,7 +237,8 @@ async fn run_tcp_tunnel(mut tunnel: ngrok::tunnel::TcpTunnel, state: DgwState) {
                         .sessions(state.sessions)
                         .subscriber_tx(state.subscriber_tx)
                         .active_recordings(state.recordings.active_recordings)
-                        .credentials(state.credentials)
+                        .provisioning(state.provisioning)
+                        .synthetic_kdc_registry(state.synthetic_kdc_registry)
                         .agent_tunnel_handle(state.agent_tunnel_handle)
                         .build()
                         .serve()
