--- devolutions-gateway/src/rdp_proxy.rs
@@ -1,846 +1,0 @@
-use std::net::{IpAddr, SocketAddr};
-use std::sync::Arc;
-
-use anyhow::Context as _;
-use ironrdp_acceptor::credssp::CredsspProcessGenerator as CredsspServerProcessGenerator;
-use ironrdp_connector::credssp::CredsspProcessGenerator as CredsspClientProcessGenerator;
-use ironrdp_connector::sspi;
-use ironrdp_connector::sspi::generator::GeneratorState;
-use ironrdp_pdu::{mcs, nego, x224};
-use secrecy::ExposeSecret as _;
-use tokio::io::{AsyncRead, AsyncWrite, AsyncWriteExt};
-use typed_builder::TypedBuilder;
-
-use crate::config::Conf;
-use crate::credential::AppCredential;
-use crate::credential_injection_kdc::{
-    CredentialInjectionClientAcceptorProtocol, CredentialInjectionKdc, CredentialInjectionKdcInterception,
-};
-use crate::kdc_connector::KdcConnector;
-use crate::proxy::Proxy;
-use crate::session::{DisconnectInterest, SessionInfo, SessionMessageSender};
-use crate::subscriber::SubscriberSender;
-
-#[derive(TypedBuilder)]
-pub struct RdpProxy<C, S> {
-    conf: Arc<Conf>,
-    session_info: SessionInfo,
-    client_stream: C,
-    client_addr: SocketAddr,
-    server_stream: S,
-    server_addr: SocketAddr,
-    credential_injection_kdc: CredentialInjectionKdc,
-    client_stream_leftover_bytes: bytes::BytesMut,
-    sessions: SessionMessageSender,
-    subscriber_tx: SubscriberSender,
-    server_dns_name: String,
-    disconnect_interest: Option<DisconnectInterest>,
-    /// Outbound dispatcher for CredSSP-originated KDC traffic. Encapsulates whether KDC
-    /// requests should attempt agent-tunnel routing (and any `jet_agent_id` pin from the
-    /// parent association token) or always go direct.
-    kdc_connector: KdcConnector,
-}
-
-impl<A, B> RdpProxy<A, B>
-where
-    A: AsyncWrite + AsyncRead + Unpin + Send,
-    B: AsyncWrite + AsyncRead + Unpin + Send,
-{
-    pub async fn run(self) -> anyhow::Result<()> {
-        handle(self).await
-    }
-}
-
-#[instrument("rdp_proxy", skip_all, fields(session_id = proxy.session_info.id.to_string(), target = proxy.server_addr.to_string()))]
-async fn handle<C, S>(proxy: RdpProxy<C, S>) -> anyhow::Result<()>
-where
-    C: AsyncRead + AsyncWrite + Unpin + Send,
-    S: AsyncRead + AsyncWrite + Unpin + Send,
-{
-    let RdpProxy {
-        conf,
-        session_info,
-        client_stream,
-        client_addr,
-        server_stream,
-        server_addr,
-        credential_injection_kdc,
-        client_stream_leftover_bytes,
-        sessions,
-        subscriber_tx,
-        server_dns_name,
-        disconnect_interest,
-        kdc_connector,
-    } = proxy;
-
-    let tls_conf = conf.credssp_tls.get().context("CredSSP TLS configuration")?;
-    let gateway_hostname = conf.hostname.clone();
-
-    // -- Retrieve the Gateway TLS public key that must be used for client-proxy CredSSP later on -- //
-
-    let gateway_cert_chain_handle = tokio::spawn(crate::tls::get_cert_chain_for_acceptor_cached(
-        gateway_hostname.clone(),
-        tls_conf.acceptor.clone(),
-    ));
-
-    // -- Dual handshake with the client and the server until the TLS security upgrade -- //
-
-    let mut client_framed =
-        ironrdp_tokio::MovableTokioFramed::new_with_leftover(client_stream, client_stream_leftover_bytes);
-    let mut server_framed = ironrdp_tokio::MovableTokioFramed::new(server_stream);
-
-    let handshake_result = dual_handshake_until_tls_upgrade(
-        &mut client_framed,
-        &mut server_framed,
-        credential_injection_kdc.target_credential(),
-    )
-    .await?;
-
-    let client_stream = client_framed.into_inner_no_leftover();
-    let server_stream = server_framed.into_inner_no_leftover();
-
-    // -- Perform the TLS upgrading for both the client and the server, effectively acting as a man-in-the-middle -- //
-
-    let client_tls_upgrade_fut = tls_conf.acceptor.accept(client_stream);
-    let server_tls_upgrade_fut = crate::tls::dangerous_connect(server_dns_name.clone(), server_stream);
-
-    let (client_stream, server_stream) = tokio::join!(client_tls_upgrade_fut, server_tls_upgrade_fut);
-
-    let client_stream = client_stream.context("TLS upgrade with client failed")?;
-    let server_stream = server_stream.context("TLS upgrade with server failed")?;
-
-    let server_public_key =
-        crate::tls::extract_stream_peer_public_key(&server_stream).context("extract target server TLS public key")?;
-
-    let gateway_cert_chain = gateway_cert_chain_handle.await??;
-    let gateway_public_key = crate::tls::extract_public_key(gateway_cert_chain.first().context("no leaf")?)
-        .context("extract Gateway public key")?;
-
-    // -- Perform the CredSSP authentication with the client (acting as a server) and the server (acting as a client) -- //
-
-    let mut client_framed = ironrdp_tokio::MovableTokioFramed::new(client_stream);
-    let mut server_framed = ironrdp_tokio::MovableTokioFramed::new(server_stream);
-
-    let krb_configs =
-        credential_injection_kerberos_configs(&conf, client_addr, &gateway_hostname, &credential_injection_kdc)?;
-
-    let client_credssp_fut = perform_credssp_as_server(
-        &mut client_framed,
-        client_addr.ip(),
-        gateway_public_key,
-        handshake_result.client_security_protocol,
-        credential_injection_kdc.proxy_credential(),
-        krb_configs.server,
-        &credential_injection_kdc,
-        &kdc_connector,
-    );
-
-    let server_credssp_fut = perform_credssp_as_client(
-        &mut server_framed,
-        server_dns_name,
-        server_public_key,
-        handshake_result.server_security_protocol,
-        credential_injection_kdc.target_credential(),
-        krb_configs.client,
-        &kdc_connector,
-    );
-
-    let (client_credssp_res, server_credssp_res) = tokio::join!(client_credssp_fut, server_credssp_fut);
-    client_credssp_res.context("CredSSP with client")?;
-    server_credssp_res.context("CredSSP with server")?;
-
-    // -- Intercept the Connect Confirm PDU, to override the server_security_protocol field -- //
-
-    intercept_connect_confirm(
-        &mut client_framed,
-        &mut server_framed,
-        handshake_result.server_security_protocol,
-    )
-    .await?;
-
-    let (mut client_stream, client_leftover) = client_framed.into_inner();
-    let (mut server_stream, server_leftover) = server_framed.into_inner();
-
-    // -- At this point, proceed to the usual two-way forwarding -- //
-
-    info!("RDP-TLS forwarding (credential injection)");
-
-    client_stream
-        .write_all(&server_leftover)
-        .await
-        .context("write server leftover to client")?;
-
-    server_stream
-        .write_all(&client_leftover)
-        .await
-        .context("write client leftover to server")?;
-
-    Proxy::builder()
-        .conf(conf)
-        .session_info(session_info)
-        .address_a(client_addr)
-        .transport_a(client_stream)
-        .address_b(server_addr)
-        .transport_b(server_stream)
-        .sessions(sessions)
-        .subscriber_tx(subscriber_tx)
-        .disconnect_interest(disconnect_interest)
-        .build()
-        .select_dissector_and_forward()
-        .await
-        .context("RDP-TLS traffic proxying failed")?;
-
-    Ok(())
-}
-
-#[derive(Debug)]
-struct HandshakeResult {
-    client_security_protocol: nego::SecurityProtocol,
-    server_security_protocol: nego::SecurityProtocol,
-}
-
-#[instrument(level = "debug", ret, skip_all)]
-pub(crate) async fn intercept_connect_confirm<C, S>(
-    client_framed: &mut ironrdp_tokio::MovableTokioFramed<C>,
-    server_framed: &mut ironrdp_tokio::MovableTokioFramed<S>,
-    server_security_protocol: nego::SecurityProtocol,
-) -> anyhow::Result<()>
-where
-    C: AsyncWrite + AsyncRead + Unpin + Send,
-    S: AsyncWrite + AsyncRead + Unpin + Send,
-{
-    let (_, received_frame) = client_framed
-        .read_pdu()
-        .await
-        .context("read MCS Connect Initial from client")?;
-    let received_connect_initial: x224::X224<x224::X224Data<'_>> =
-        ironrdp_core::decode(&received_frame).context("decode PDU from client")?;
-    let mut received_connect_initial: mcs::ConnectInitial =
-        ironrdp_core::decode(&received_connect_initial.0.data).context("decode Connect Initial PDU")?;
-    trace!(message = ?received_connect_initial, "Received Connect Initial PDU from client");
-
-    let mut gcc_blocks = received_connect_initial.conference_create_request.into_gcc_blocks();
-    gcc_blocks.core.optional_data.server_selected_protocol = Some(server_security_protocol);
-    // Update the conference request with modified gcc_blocks.
-    received_connect_initial.conference_create_request = ironrdp_pdu::gcc::ConferenceCreateRequest::new(gcc_blocks)?;
-    trace!(message = ?received_connect_initial, "Send Connection Request PDU to server");
-    let x224_msg_buf = ironrdp_core::encode_vec(&received_connect_initial)?;
-    let pdu = x224::X224Data {
-        data: std::borrow::Cow::Owned(x224_msg_buf),
-    };
-    send_pdu(server_framed, &x224::X224(pdu))
-        .await
-        .context("send connection request to server")?;
-
-    Ok(())
-}
-
-#[instrument(name = "dual_handshake", level = "debug", ret, skip_all)]
-async fn dual_handshake_until_tls_upgrade<C, S>(
-    client_framed: &mut ironrdp_tokio::MovableTokioFramed<C>,
-    server_framed: &mut ironrdp_tokio::MovableTokioFramed<S>,
-    target_credential: &AppCredential,
-) -> anyhow::Result<HandshakeResult>
-where
-    C: AsyncWrite + AsyncRead + Unpin + Send,
-    S: AsyncWrite + AsyncRead + Unpin + Send,
-{
-    let (_, received_frame) = client_framed.read_pdu().await.context("read PDU from client")?;
-    let received_connection_request: x224::X224<nego::ConnectionRequest> =
-        ironrdp_core::decode(&received_frame).context("decode PDU from client")?;
-    trace!(message = ?received_connection_request, "Received Connection Request PDU from client");
-
-    // Choose the security protocol to use with the client.
-    let received_connection_request_protocol = received_connection_request.0.protocol;
-    let client_security_protocol = if received_connection_request_protocol.contains(nego::SecurityProtocol::HYBRID_EX) {
-        nego::SecurityProtocol::HYBRID_EX
-    } else if received_connection_request
-        .0
-        .protocol
-        .contains(nego::SecurityProtocol::HYBRID)
-    {
-        nego::SecurityProtocol::HYBRID
-    } else {
-        anyhow::bail!(
-            "client does not support CredSSP (received {})",
-            received_connection_request.0.protocol
-        )
-    };
-
-    let connection_request_to_send = nego::ConnectionRequest {
-        nego_data: match target_credential {
-            AppCredential::UsernamePassword { username, .. } => {
-                Some(nego::NegoRequestData::cookie(username.to_owned()))
-            }
-        },
-        flags: received_connection_request.0.flags,
-        // https://learn.microsoft.com/en-us/openspecs/windows_protocols/ms-rdpbcgr/902b090b-9cb3-4efc-92bf-ee13373371e3
-        //
-        // The spec states that `PROTOCOL_SSL` "SHOULD" also be set when using `PROTOCOL_HYBRID`:
-        //
-        // > PROTOCOL_HYBRID (0x00000002)
-        // > Credential Security Support Provider protocol (CredSSP) (section 5.4.5.2).
-        // > If this flag is set, then the PROTOCOL_SSL (0x00000001) flag SHOULD also be set
-        // > because Transport Layer Security (TLS) is a subset of CredSSP.
-        //
-        // However, in practice `mstsc` is picky about these flags: it expects the
-        // SupportedProtocol bits in the ConnectionRequestPDU that reach the target
-        // server to match what the client originally sent. If the proxy modifies
-        // them (for example, forcing HYBRID | HYBRID_EX and/or clearing SSL),
-        // the connection can fail with an authentication error (Code: 0x609).
-        //
-        // We therefore *do not* synthesize a new protocol bitmask here anymore.
-        // Instead, we forward the client's SupportedProtocol flags as-is and
-        // enforce our policy by validating them: if HYBRID / HYBRID_EX are not
-        // present (i.e. NLA is not negotiated), we fail the connection rather
-        // than trying to "fix" the flags ourselves.
-        //
-        // See also: https://serverfault.com/a/720161
-        protocol: received_connection_request_protocol,
-    };
-    trace!(?connection_request_to_send, "Send Connection Request PDU to server");
-    send_pdu(server_framed, &x224::X224(connection_request_to_send))
-        .await
-        .context("send connection request to server")?;
-
-    let (_, received_frame) = server_framed.read_pdu().await.context("read PDU from server")?;
-    let received_connection_confirm: x224::X224<nego::ConnectionConfirm> =
-        ironrdp_core::decode(&received_frame).context("decode PDU from server")?;
-    trace!(message = ?received_connection_confirm, "Received Connection Confirm PDU from server");
-
-    let (connection_confirm_to_send, handshake_result) = match &received_connection_confirm.0 {
-        nego::ConnectionConfirm::Response {
-            flags,
-            protocol: server_security_protocol,
-        } => {
-            debug!(?server_security_protocol, ?flags, "Server confirmed connection");
-
-            let result = if !server_security_protocol
-                .intersects(nego::SecurityProtocol::HYBRID | nego::SecurityProtocol::HYBRID_EX)
-            {
-                Err(anyhow::anyhow!(
-                    "server selected security protocol {server_security_protocol}, which is not supported for credential injection"
-                ))
-            } else {
-                Ok(HandshakeResult {
-                    client_security_protocol,
-                    server_security_protocol: *server_security_protocol,
-                })
-            };
-
-            (
-                x224::X224(nego::ConnectionConfirm::Response {
-                    flags: *flags,
-                    protocol: client_security_protocol,
-                }),
-                result,
-            )
-        }
-        nego::ConnectionConfirm::Failure { code } => (
-            x224::X224(received_connection_confirm.0.clone()),
-            Err(anyhow::anyhow!("RDP session initiation failed with code {code}")),
-        ),
-    };
-
-    trace!(?connection_confirm_to_send, "Send Connection Request PDU to client");
-    send_pdu(client_framed, &connection_confirm_to_send)
-        .await
-        .context("send connection confirm to client")?;
-
-    handshake_result
-}
-
-/// Kerberos configs for the two CredSSP legs of a credential-injection session.
-///
-/// `server` drives the client-facing acceptor (Gateway-as-server); `client` drives the
-/// target-facing leg (Gateway-as-client). `None` on a leg means that leg authenticates over NTLM.
-pub(crate) struct CredentialInjectionKerberosConfigs {
-    pub server: Option<sspi::KerberosServerConfig>,
-    pub client: Option<ironrdp_connector::credssp::KerberosConfig>,
-}
-
-/// Whether a credential-injection session speaks Kerberos (vs NTLM). Decided once so both CredSSP
-/// legs agree — sspi's acceptor and initiator must speak the same package or the handshake fails
-/// reading one as the other. Kerberos needs the experimental opt-in AND a domain-qualified target
-/// (a domainless account can't get a ticket).
-fn injection_uses_kerberos(
-    enable_unstable: bool,
-    kerberos_credential_injection: bool,
-    protocol: CredentialInjectionClientAcceptorProtocol,
-) -> bool {
-    enable_unstable
-        && kerberos_credential_injection
-        && matches!(protocol, CredentialInjectionClientAcceptorProtocol::Kerberos)
-}
-
-/// Build the Kerberos config for both CredSSP legs from the single [`injection_uses_kerberos`]
-/// decision. Everything else is NTLM on both legs.
-pub(crate) fn credential_injection_kerberos_configs(
-    conf: &Conf,
-    client_addr: SocketAddr,
-    gateway_hostname: &str,
-    credential_injection_kdc: &CredentialInjectionKdc,
-) -> anyhow::Result<CredentialInjectionKerberosConfigs> {
-    let protocol = credential_injection_kdc.client_acceptor_protocol()?;
-
-    if !injection_uses_kerberos(
-        conf.debug.enable_unstable,
-        conf.debug.kerberos_credential_injection,
-        protocol,
-    ) {
-        return Ok(CredentialInjectionKerberosConfigs {
-            server: None,
-            client: None,
-        });
-    }
-
-    let krb_kdc = credential_injection_kdc
-        .krb_kdc()
-        .context("kerberos credential injection requires the krb_kdc target connection option")?;
-
-    Ok(CredentialInjectionKerberosConfigs {
-        server: Some(credential_injection_kdc.server_kerberos_config(client_addr)?),
-        client: Some(ironrdp_connector::credssp::KerberosConfig {
-            kdc_proxy_url: Some(krb_kdc.clone()),
-            hostname: gateway_hostname.to_owned(),
-        }),
-    })
-}
-
-#[instrument(name = "server_credssp", level = "debug", ret, skip_all)]
-pub(crate) async fn perform_credssp_as_client<S>(
-    framed: &mut ironrdp_tokio::Framed<S>,
-    server_name: String,
-    server_public_key: Vec<u8>,
-    security_protocol: nego::SecurityProtocol,
-    credentials: &AppCredential,
-    kerberos_config: Option<ironrdp_connector::credssp::KerberosConfig>,
-    kdc_connector: &KdcConnector,
-) -> anyhow::Result<()>
-where
-    S: ironrdp_tokio::FramedRead + ironrdp_tokio::FramedWrite,
-{
-    use ironrdp_tokio::FramedWrite as _;
-
-    // Decrypt password into short-lived buffer.
-    let (username, decrypted_password) = credentials
-        .decrypt_password()
-        .context("failed to decrypt credentials")?;
-
-    let credentials = ironrdp_connector::Credentials::UsernamePassword {
-        username,
-        password: decrypted_password.expose_secret().to_owned(),
-    };
-    // decrypted_password drops here, zeroizing its buffer; note: a copy of the plaintext
-    // remains in `credentials` above, which is a regular String (downstream API limitation).
-
-    let (mut sequence, mut ts_request) = ironrdp_connector::credssp::CredsspSequence::init(
-        credentials,
-        None,
-        security_protocol,
-        ironrdp_connector::ServerName::new(server_name.clone()),
-        server_public_key,
-        kerberos_config,
-    )?;
-
-    let mut buf = ironrdp_pdu::WriteBuf::new();
-
-    loop {
-        let client_state = {
-            let mut generator = sequence.process_ts_request(ts_request);
-            resolve_client_generator(&mut generator, kdc_connector).await?
-        }; // drop generator
-
-        buf.clear();
-        let written = sequence.handle_process_result(client_state, &mut buf)?;
-
-        if let Some(response_len) = written.size() {
-            let response = &buf[..response_len];
-            framed
-                .write_all(response)
-                .await
-                .map_err(|e| ironrdp_connector::custom_err!("write all", e))?;
-        }
-
-        let Some(next_pdu_hint) = sequence.next_pdu_hint() else {
-            break;
-        };
-
-        let pdu = framed.read_by_hint(next_pdu_hint).await.context("read frame by hint")?;
-
-        if let Some(next_request) = sequence.decode_server_message(&pdu)? {
-            ts_request = next_request;
-        } else {
-            break;
-        }
-    }
-
-    Ok(())
-}
-
-async fn resolve_server_generator(
-    generator: &mut CredsspServerProcessGenerator<'_>,
-    credential_injection_kdc: &CredentialInjectionKdc,
-    kdc_connector: &KdcConnector,
-) -> Result<sspi::credssp::ServerState, sspi::credssp::ServerError> {
-    let mut state = generator.start();
-
-    loop {
-        match state {
-            GeneratorState::Suspended(request) => {
-                let response = match credential_injection_kdc.intercept_network_request(&request) {
-                    Ok(CredentialInjectionKdcInterception::Intercepted(response)) => Ok(response),
-                    Ok(CredentialInjectionKdcInterception::NotInjectionRequest) => {
-                        kdc_connector.send_network_request(&request).await
-                    }
-                    Ok(CredentialInjectionKdcInterception::NotInjectionRealm(mismatch)) => Err(anyhow::anyhow!(
-                        "kdc request realm does not match credential-injection session realm: {mismatch}"
-                    )),
-                    Err(error) => Err(error),
-                }
-                .map_err(|err| sspi::credssp::ServerError {
-                    ts_request: None,
-                    error: sspi::Error::new(sspi::ErrorKind::InternalError, err),
-                })?;
-
-                state = generator.resume(Ok(response));
-            }
-            GeneratorState::Completed(client_state) => {
-                break client_state;
-            }
-        }
-    }
-}
-
-async fn resolve_client_generator(
-    generator: &mut CredsspClientProcessGenerator<'_>,
-    kdc_connector: &KdcConnector,
-) -> anyhow::Result<sspi::credssp::ClientState> {
-    let mut state = generator.start();
-
-    loop {
-        match state {
-            GeneratorState::Suspended(request) => {
-                let response = kdc_connector.send_network_request(&request).await?;
-                state = generator.resume(Ok(response));
-            }
-            GeneratorState::Completed(client_state) => {
-                break Ok(client_state.map_err(|e| {
-                    ironrdp_connector::ConnectorError::new("CredSSP", ironrdp_connector::ConnectorErrorKind::Credssp(e))
-                })?);
-            }
-        };
-    }
-}
-
-#[expect(clippy::too_many_arguments)]
-#[instrument(name = "client_credssp", level = "debug", ret, skip_all)]
-pub(crate) async fn perform_credssp_as_server<S>(
-    framed: &mut ironrdp_tokio::Framed<S>,
-    client_addr: IpAddr,
-    gateway_public_key: Vec<u8>,
-    security_protocol: nego::SecurityProtocol,
-    credentials: &AppCredential,
-    kerberos_server_config: Option<sspi::KerberosServerConfig>,
-    credential_injection_kdc: &CredentialInjectionKdc,
-    kdc_connector: &KdcConnector,
-) -> anyhow::Result<()>
-where
-    S: ironrdp_tokio::FramedRead + ironrdp_tokio::FramedWrite,
-{
-    use ironrdp_connector::sspi::credssp::EarlyUserAuthResult;
-    use ironrdp_tokio::FramedWrite as _;
-
-    let mut buf = ironrdp_pdu::WriteBuf::new();
-
-    // Are we supposed to use the actual computer name of the client?
-    // But this does not seem to matter so far, so we stringify the IP address of the client instead.
-    let client_computer_name = ironrdp_connector::ServerName::new(client_addr.to_string());
-
-    let result = credssp_loop(
-        framed,
-        &mut buf,
-        client_computer_name,
-        gateway_public_key,
-        credentials,
-        kerberos_server_config,
-        credential_injection_kdc,
-        kdc_connector,
-    )
-    .await;
-
-    if security_protocol.intersects(nego::SecurityProtocol::HYBRID_EX) {
-        trace!(?result, "HYBRID_EX");
-
-        let result = if result.is_ok() {
-            EarlyUserAuthResult::Success
-        } else {
-            EarlyUserAuthResult::AccessDenied
-        };
-
-        buf.clear();
-        result.to_buffer(&mut buf).context("write early user auth result")?;
-        let response = &buf[..result.buffer_len()];
-        framed.write_all(response).await.context("write_all")?;
-    }
-
-    return result;
-
-    async fn credssp_loop<S>(
-        framed: &mut ironrdp_tokio::Framed<S>,
-        buf: &mut ironrdp_pdu::WriteBuf,
-        client_computer_name: ironrdp_connector::ServerName,
-        public_key: Vec<u8>,
-        credentials: &AppCredential,
-        kerberos_server_config: Option<sspi::KerberosServerConfig>,
-        credential_injection_kdc: &CredentialInjectionKdc,
-        kdc_connector: &KdcConnector,
-    ) -> anyhow::Result<()>
-    where
-        S: ironrdp_tokio::FramedRead + ironrdp_tokio::FramedWrite,
-    {
-        // Decrypt password into short-lived buffer.
-        let (username, decrypted_password) = credentials
-            .decrypt_password()
-            .context("failed to decrypt credentials")?;
-
-        let username = sspi::Username::parse(&username).context("invalid username")?;
-
-        let identity = sspi::AuthIdentity {
-            username,
-            password: decrypted_password.expose_secret().to_owned().into(),
-        };
-        // decrypted_password drops here, zeroizing its buffer; note: a copy of the plaintext
-        // remains in `identity` above (downstream API limitation).
-
-        let mut sequence = ironrdp_acceptor::credssp::CredsspSequence::init(
-            &identity,
-            client_computer_name,
-            public_key,
-            kerberos_server_config,
-        )?;
-
-        loop {
-            let Some(next_pdu_hint) = sequence.next_pdu_hint()? else {
-                break;
-            };
-
-            let pdu = framed
-                .read_by_hint(next_pdu_hint)
-                .await
-                .map_err(|e| ironrdp_connector::custom_err!("read frame by hint", e))?;
-
-            let Some(ts_request) = sequence.decode_client_message(&pdu)? else {
-                break;
-            };
-
-            let result = {
-                let mut generator = sequence.process_ts_request(ts_request);
-                resolve_server_generator(&mut generator, credential_injection_kdc, kdc_connector).await
-            }; // drop generator
-
-            buf.clear();
-            let written = sequence.handle_process_result(result, buf)?;
-
-            if let Some(response_len) = written.size() {
-                let response = &buf[..response_len];
-                framed
-                    .write_all(response)
-                    .await
-                    .map_err(|e| ironrdp_connector::custom_err!("write all", e))?;
-            }
-        }
-
-        Ok(())
-    }
-}
-
-async fn send_pdu<S, P>(framed: &mut ironrdp_tokio::MovableTokioFramed<S>, pdu: &P) -> anyhow::Result<()>
-where
-    S: AsyncWrite + Unpin + Send,
-    P: ironrdp_core::Encode,
-{
-    use ironrdp_tokio::FramedWrite as _;
-
-    let payload = ironrdp_core::encode_vec(pdu).context("failed to encode PDU")?;
-    framed.write_all(&payload).await.context("failed to write PDU")?;
-    Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use std::net::{Ipv4Addr, SocketAddr};
-    use std::sync::Arc;
-
-    use base64::Engine as _;
-    use secrecy::SecretString;
-    use uuid::Uuid;
-
-    use super::*;
-    use crate::config::ConfHandle;
-    use crate::credential::{CleartextAppCredential, CleartextAppCredentialMapping};
-    use crate::credential_injection_kdc::CredentialService;
-    use crate::target_connection_options::TargetConnectionOptions;
-
-    const TEST_CONFIG: &str = r#"{
-        "Hostname": "dgateway.localhost.com",
-        "ProvisionerPublicKeyData": {
-            "Value": "mMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA4vuqLOkl1pWobt6su1XO9VskgCAwevEGs6kkNjJQBwkGnPKYLmNF1E/af1yCocfVn/OnPf9e4x+lXVyZ6LMDJxFxu+axdgOq3Ld392J1iAEbfvwlyRFnEXFOJNyylqg3bY6LvnWHL/XZczVdMD9xYfq2sO9bg3xjRW4s7r9EEYOFjqVT3VFznH9iWJVtcSEKukmS/3uKoO6lGhacvu0HhjXXdgq0R8zvR4XRJ9Fcnf0f9Ypoc+i6L80NVjrRCeVOH+Ld/2fA9bocpfLarcVqG3RjS+qgOtpyCc0jWVFF4zaGQ7LUDFkEIYILkICeMMn2ll29hmZNzsJzZJ9s6NocgQIDAQAB"
-        },
-        "Listeners": [
-            { "InternalUrl": "http://*:7171", "ExternalUrl": "https://*:7171" }
-        ],
-        "__debug__": { "disable_token_validation": true }
-    }"#;
-
-    const KERBEROS_CONFIG: &str = r#"{
-        "Hostname": "dgateway.localhost.com",
-        "ProvisionerPublicKeyData": {
-            "Value": "mMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA4vuqLOkl1pWobt6su1XO9VskgCAwevEGs6kkNjJQBwkGnPKYLmNF1E/af1yCocfVn/OnPf9e4x+lXVyZ6LMDJxFxu+axdgOq3Ld392J1iAEbfvwlyRFnEXFOJNyylqg3bY6LvnWHL/XZczVdMD9xYfq2sO9bg3xjRW4s7r9EEYOFjqVT3VFznH9iWJVtcSEKukmS/3uKoO6lGhacvu0HhjXXdgq0R8zvR4XRJ9Fcnf0f9Ypoc+i6L80NVjrRCeVOH+Ld/2fA9bocpfLarcVqG3RjS+qgOtpyCc0jWVFF4zaGQ7LUDFkEIYILkICeMMn2ll29hmZNzsJzZJ9s6NocgQIDAQAB"
-        },
-        "Listeners": [
-            { "InternalUrl": "http://*:7171", "ExternalUrl": "https://*:7171" }
-        ],
-        "__debug__": {
-            "disable_token_validation": true,
-            "enable_unstable": true,
-            "kerberos_credential_injection": true
-        }
-    }"#;
-
-    fn conf(json: &str) -> Arc<Conf> {
-        ConfHandle::mock(json).expect("test config is valid").get_conf()
-    }
-
-    fn client_addr() -> SocketAddr {
-        SocketAddr::from((Ipv4Addr::LOCALHOST, 33_889))
-    }
-
-    fn association_token(jti: Uuid) -> String {
-        let engine = base64::engine::general_purpose::URL_SAFE_NO_PAD;
-        let header = engine.encode(r#"{"alg":"RS256"}"#);
-        let payload = engine.encode(
-            serde_json::to_vec(&serde_json::json!({
-                "jti": jti,
-                "dst_hst": "target.example:3389"
-            }))
-            .expect("payload serializes"),
-        );
-        let signature = engine.encode(b"signature");
-        format!("{header}.{payload}.{signature}")
-    }
-
-    fn mapping(target_username: &str) -> CleartextAppCredentialMapping {
-        CleartextAppCredentialMapping {
-            proxy: CleartextAppCredential::UsernamePassword {
-                username: "proxy@example.invalid".to_owned(),
-                password: SecretString::from("pwd"),
-            },
-            target: CleartextAppCredential::UsernamePassword {
-                username: target_username.to_owned(),
-                password: SecretString::from("pwd"),
-            },
-        }
-    }
-
-    /// Provision credentials (and optional `krb_kdc`) then resolve the injection KDC — the
-    /// in-process path RDP takes before building CredSSP Kerberos configs.
-    fn provisioned_kdc(target_username: &str, krb_kdc: Option<&str>) -> CredentialInjectionKdc {
-        let service = CredentialService::new(ConfHandle::mock(TEST_CONFIG).expect("test config is valid"));
-        let jti = Uuid::new_v4();
-        service
-            .insert_credentials(
-                association_token(jti),
-                Some(mapping(target_username)),
-                time::Duration::minutes(5),
-            )
-            .expect("credentials insert");
-        if let Some(krb_kdc) = krb_kdc {
-            let options = TargetConnectionOptions::new(Some(krb_kdc)).expect("valid krb_kdc");
-            service.insert_connection_options(jti, options, time::Duration::minutes(5));
-        }
-        service.kdc_for(jti).expect("kdc_for resolves provisioned state")
-    }
-
-    // The two CredSSP legs are built from this single decision, so agreement is guaranteed by
-    // construction. These cases pin the decision itself (the bug was the two legs deciding
-    // independently): Kerberos requires BOTH opt-in flags AND a domain-qualified target.
-    #[test]
-    fn injection_uses_kerberos_requires_optin_and_domain_qualified_target() {
-        use CredentialInjectionClientAcceptorProtocol::{Kerberos, Ntlm};
-
-        assert!(injection_uses_kerberos(true, true, Kerberos));
-
-        // Either opt-in off => NTLM, even for a Kerberos-capable target.
-        assert!(!injection_uses_kerberos(false, true, Kerberos));
-        assert!(!injection_uses_kerberos(true, false, Kerberos));
-
-        // Domainless target can't get a ticket => NTLM regardless of the flags.
-        assert!(!injection_uses_kerberos(true, true, Ntlm));
-        assert!(!injection_uses_kerberos(false, false, Ntlm));
-    }
-
-    #[test]
-    fn provisioned_krb_kdc_becomes_client_kdc_proxy_url() {
-        let conf = conf(KERBEROS_CONFIG);
-        let kdc = provisioned_kdc("administrator@example.invalid", Some("tcp://dc.example.com:88"));
-
-        let configs =
-            credential_injection_kerberos_configs(conf.as_ref(), client_addr(), "dgateway.localhost.com", &kdc)
-                .expect("kerberos configs build when krb_kdc is provisioned");
-
-        let client = configs.client.expect("client leg speaks Kerberos");
-        assert_eq!(
-            client.kdc_proxy_url.as_ref().map(url::Url::as_str),
-            Some("tcp://dc.example.com:88"),
-            "target-side CredSSP must use the provisioned KDC URL",
-        );
-        assert_eq!(client.hostname, "dgateway.localhost.com");
-        assert!(configs.server.is_some(), "both CredSSP legs must agree on Kerberos");
-    }
-
-    #[test]
-    fn kerberos_path_requires_provisioned_krb_kdc() {
-        let conf = conf(KERBEROS_CONFIG);
-        let kdc = provisioned_kdc("administrator@example.invalid", None);
-
-        let error =
-            match credential_injection_kerberos_configs(conf.as_ref(), client_addr(), "dgateway.localhost.com", &kdc) {
-                Ok(_) => panic!("Kerberos without krb_kdc must fail before CredSSP starts"),
-                Err(error) => error,
-            };
-
-        assert!(
-            format!("{error:#}").contains("krb_kdc"),
-            "error should name the missing connection option, got: {error:#}",
-        );
-    }
-
-    #[test]
-    fn ntlm_path_does_not_require_krb_kdc() {
-        // Domainless target → NTLM decision even with Kerberos feature flags on.
-        let conf = conf(KERBEROS_CONFIG);
-        let kdc = provisioned_kdc("Administrator", None);
-
-        let configs =
-            credential_injection_kerberos_configs(conf.as_ref(), client_addr(), "dgateway.localhost.com", &kdc)
-                .expect("NTLM path succeeds without connection options");
-
-        assert!(configs.client.is_none());
-        assert!(configs.server.is_none());
-    }
-
-    #[test]
-    fn kerberos_flags_off_does_not_require_krb_kdc() {
-        // Domain-qualified target but feature flags off → NTLM on both legs.
-        let conf = conf(TEST_CONFIG);
-        let kdc = provisioned_kdc("administrator@example.invalid", None);
-
-        let configs =
-            credential_injection_kerberos_configs(conf.as_ref(), client_addr(), "dgateway.localhost.com", &kdc)
-                .expect("flags off means NTLM without needing krb_kdc");
-
-        assert!(configs.client.is_none());
-        assert!(configs.server.is_none());
-    }
-}
